Add files via upload
Add initial files for simple chatbot deployment

Added app.py file with the Flask application code for the chatbot, index.html file in the templates folder for the user interface, and style.css file in the static folder for styling. Also added a Procfile to specify the command to start the app on Heroku and a requirements.txt file with the necessary dependencies. The database file db.sqlite3 and its supporting files were also added to store conversation data.